Cache in main file and no-ops when there is no cache
